feat(cmd): add project and docker subcommands to scan
